100-get_endianness.c

+++ 0x14-bit_manipulation/test/100-get_endianness.c
@@ -1,13 +1,13 @@
 #include "main.h"
 
 /**
- * get_endianness - Checks if computer is big or little endian
- * Return: 0 if big edian and 1 if little edian
+ * get_endianness -  checks the endianness
+ * Return: 0 if big endian, 1 if little endian
  */
 int get_endianness(void)
 {
 	unsigned int i = 1;
 	char *c = (char *) &i;
 
-	return (*c);
+	return ((int)*c);
 }
